musl: Use long in read syscall

+++ src/musl/read.cpp
@@ -1,7 +1,7 @@
 #include "common.hpp"
 #include <fd_map.hpp>
 
-static ssize_t sys_read(int fd, void* buf, size_t len) {
+static long sys_read(int fd, void* buf, size_t len) {
   try {
     auto& fildes = FD_map::_get(fd);
     return fildes.read(buf, len);
@@ -12,6 +12,6 @@ static ssize_t sys_read(int fd, void* buf, size_t len) {
 }
 
 extern "C"
-ssize_t syscall_SYS_read(int fd, void *buf, size_t nbyte) {
+long syscall_SYS_read(int fd, void *buf, size_t nbyte) {
   return strace(sys_read, "read", fd, buf, nbyte);
 }
